+++ example-projects/java17/src/main/java/module-info.java
@@ -1,4 +1,0 @@
-module jgiven.exampleprojects.java17 {
-    exports com.tngtech.jgiven.exampleprojects.java17;
-    requires java.base;
-}